# Fix performance bottleneck on benchmark