I tried to cleanup benchmark code, and shave some overhead that is not related to data structures.
 - Clean up benchmark code with macros and consts
 - Use XorShiftRng instead of thread_rng, which is faster and deterministic.
 - Change String to Vec<u8>, which should give same results for each data structures. Constructing Vec<u8> is much cheaper than String because Rust String needs an extra calculation UTF-8 validation. Also uses with_capacity() instead of collect(), which reduces memory allocation caused by buffer resizing.

before
```
$ cargo bench
   Compiling art v0.1.0 (file:///home/yjh0502/repo/git/ArtTree)
    Finished release [optimized] target(s) in 6.2 secs
     Running target/release/deps/art-a315e05db900bbaf

running 21 tests
test bench::bench::bench_insert_art_len_1000_rnd_string   ... bench:  42,450,417 ns/iter (+/- 1,557,903)
test bench::bench::bench_insert_art_len_100_rnd_string    ... bench:  34,051,892 ns/iter (+/- 796,406)
test bench::bench::bench_insert_art_len_20_rnd_string     ... bench:  54,533,373 ns/iter (+/- 1,002,201)
test bench::bench::bench_insert_art_u32                   ... bench:  26,527,038 ns/iter (+/- 1,806,945)
test bench::bench::bench_insert_art_u64                   ... bench:  30,681,753 ns/iter (+/- 1,129,419)
test bench::bench::bench_insert_btree_len_1000_rnd_string ... bench:  41,386,161 ns/iter (+/- 111,107)
test bench::bench::bench_insert_btree_len_100_rnd_string  ... bench:  35,947,975 ns/iter (+/- 277,836)
test bench::bench::bench_insert_btree_len_20_rnd_string   ... bench:  62,120,743 ns/iter (+/- 685,137)
test bench::bench::bench_insert_btree_u32                 ... bench:  23,803,864 ns/iter (+/- 183,533)
test bench::bench::bench_insert_btree_u64                 ... bench:  28,162,813 ns/iter (+/- 355,500)
test bench::bench::bench_insert_hmap_len_1000_rnd_string  ... bench:  41,805,358 ns/iter (+/- 84,059)
test bench::bench::bench_insert_hmap_len_100_rnd_string   ... bench:  33,440,761 ns/iter (+/- 272,043)
test bench::bench::bench_insert_hmap_len_20_rnd_string    ... bench:  50,518,336 ns/iter (+/- 903,199)
test bench::bench::bench_insert_hmap_u32                  ... bench:  16,807,355 ns/iter (+/- 286,077)
test bench::bench::bench_insert_hmap_u64                  ... bench:  19,023,666 ns/iter (+/- 499,562)
test bench::bench::bench_search_art_rnd_u64               ... bench:  61,168,162 ns/iter (+/- 13,680,195)
test bench::bench::bench_search_art_seq_u64               ... bench:   6,261,751 ns/iter (+/- 423,354)
test bench::bench::bench_search_btree_rnd_u64             ... bench:  27,038,648 ns/iter (+/- 35,145)
test bench::bench::bench_search_btree_seq_u64             ... bench:  80,660,886 ns/iter (+/- 480,733)
test bench::bench::bench_search_hmap_rnd_u64              ... bench:  43,767,141 ns/iter (+/- 245,012)
test bench::bench::bench_search_hmap_seq_u64              ... bench:  52,985,405 ns/iter (+/- 717,343)

test result: ok. 0 passed; 0 failed; 0 ignored; 21 measured

```

after
```
$ cargo bench
    Finished release [optimized] target(s) in 0.0 secs
     Running target/release/deps/art-a315e05db900bbaf

running 21 tests
test bench::bench::bench_insert_art_1000_rnd_str   ... bench:   1,855,701 ns/iter (+/- 130,305)
test bench::bench::bench_insert_art_100_rnd_str    ... bench:   4,559,853 ns/iter (+/- 349,051)
test bench::bench::bench_insert_art_20_rnd_str     ... bench:  21,458,012 ns/iter (+/- 958,960)
test bench::bench::bench_insert_art_u32            ... bench:  22,279,996 ns/iter (+/- 4,363,856)
test bench::bench::bench_insert_art_u64            ... bench:  23,274,810 ns/iter (+/- 1,397,472)
test bench::bench::bench_insert_btree_1000_rnd_str ... bench:     585,022 ns/iter (+/- 5,371)
test bench::bench::bench_insert_btree_100_rnd_str  ... bench:   4,078,027 ns/iter (+/- 39,537)
test bench::bench::bench_insert_btree_20_rnd_str   ... bench:  29,241,780 ns/iter (+/- 410,627)
test bench::bench::bench_insert_btree_u32          ... bench:  19,871,270 ns/iter (+/- 183,885)
test bench::bench::bench_insert_btree_u64          ... bench:  21,526,949 ns/iter (+/- 216,056)
test bench::bench::bench_insert_hmap_1000_rnd_str  ... bench:     484,700 ns/iter (+/- 7,443)
test bench::bench::bench_insert_hmap_100_rnd_str   ... bench:   2,739,932 ns/iter (+/- 52,738)
test bench::bench::bench_insert_hmap_20_rnd_str    ... bench:  17,483,848 ns/iter (+/- 459,348)
test bench::bench::bench_insert_hmap_u32           ... bench:  12,361,467 ns/iter (+/- 318,740)
test bench::bench::bench_insert_hmap_u64           ... bench:  12,715,853 ns/iter (+/- 310,442)
test bench::bench::bench_search_art_rnd_u64        ... bench:  68,919,512 ns/iter (+/- 3,652,568)
test bench::bench::bench_search_art_seq_u64        ... bench:   6,993,965 ns/iter (+/- 537,857)
test bench::bench::bench_search_btree_rnd_u64      ... bench:  27,704,435 ns/iter (+/- 38,993)
test bench::bench::bench_search_btree_seq_u64      ... bench:   8,091,358 ns/iter (+/- 264,220)
test bench::bench::bench_search_hmap_rnd_u64       ... bench:  46,673,854 ns/iter (+/- 281,649)
test bench::bench::bench_search_hmap_seq_u64       ... bench:   6,144,797 ns/iter (+/- 74,108)

test result: ok. 0 passed; 0 failed; 0 ignored; 21 measured
```